+++ browser/popup_handler.py
