+++ core/ocr_utils.py
