+++ core/text_clicker.py
